# feat: Add comprehensive performance testing framework with Artillery.js