--- a/demos/typescript-api/tests/performance/data/performance-baseline.json
+++ b/demos/typescript-api/tests/performance/data/performance-baseline.json
@@ -1,0 +1,22 @@
+{
+  "timestamp": "2025-08-29T16:39:41.029Z",
+  "commit": "efb9f7a9376f29f35a3d752a602484c70fa0c03d",
+  "environment": {
+    "node_version": "v22.18.0",
+    "platform": "darwin",
+    "arch": "arm64"
+  },
+  "metrics": {
+    "p95": 0,
+    "p99": 0,
+    "error_rate": 0,
+    "rps": 0,
+    "total_requests": 0
+  },
+  "thresholds": {
+    "p95_response_time": 0,
+    "p99_response_time": 0,
+    "error_rate": 0.5,
+    "requests_per_second": 0
+  }
+}