## Summary
Introduces enterprise-grade performance testing capabilities with Artillery.js, providing comprehensive load testing, monitoring, and regression detection integrated into the progressive CI/CD strictness model.

## 🚀 Key Features

### Performance Test Suite
- **Light Load Testing** (5-10 req/s): Basic validation for feature branches
- **Moderate Load Testing** (20-50 req/s): Sustained load for develop branch
- **Heavy Load Testing** (100-200 req/s): Stress testing for main branch
- **Memory Leak Detection**: 5-minute sustained load testing
- **Custom Metrics Collection**: SLA validation and endpoint-specific metrics

### 📊 Performance Monitoring & Dashboards
- Interactive HTML dashboard with real-time metrics visualization
- Historical trend analysis using Chart.js
- Performance baseline management and automated comparison
- Regression detection with configurable thresholds (20% tolerance)
- Performance alerting for P95/P99 response times, error rates, and throughput

### 🔄 CI/CD Integration
- **Progressive Performance Testing** aligned with CI strictness:
  - Basic CI (feature/*): Light performance validation
  - Comprehensive CI (develop): Moderate load testing
  - Ultra-Strict CI (main): Heavy stress testing + regression detection
- Automated baseline recording and comparison
- Performance gate criteria with appropriate thresholds

### 📈 Enterprise Capabilities
- Artillery.js test scenarios with realistic API workflows
- Automated baseline establishment with tolerance thresholds
- HTML report generation for baseline and regression analysis
- Dashboard data updater for live performance monitoring
- Comprehensive documentation with troubleshooting guides

## ⚡ Performance Thresholds

| Test Type | P95 Threshold | P99 Threshold | Error Rate | Use Case |
|-----------|---------------|---------------|------------|----------|
| Light Load | ≤200ms | ≤500ms | ≤0.1% | Feature development |
| Moderate Load | ≤300ms | ≤800ms | ≤1.0% | Staging validation |
| Heavy Load | ≤500ms | ≤1500ms | ≤5.0% | Production readiness |

## 🛠️ Implementation Details

### File Structure
```
tests/performance/
├── scenarios/           # Artillery.js test scenarios
├── scripts/            # Performance utilities & processors
├── dashboard/          # Interactive monitoring UI
├── data/              # Performance data storage
├── reports/           # Generated HTML reports
└── README.md          # Comprehensive documentation
```

### New npm Scripts
- `npm run test:perf:light` - Light load testing
- `npm run test:perf:moderate` - Moderate load testing  
- `npm run test:perf:heavy` - Heavy stress testing
- `npm run test:perf:baseline` - Record performance baseline
- `npm run test:perf:regression` - Detect performance regressions
- `npm run perf:dashboard` - Generate performance dashboard

## 🧪 Test Plan
- [x] Light performance test execution
- [x] Baseline recording and validation
- [x] Regression detection functionality
- [x] Dashboard generation and data visualization
- [x] Linting and type checking validation
- [x] All performance scripts functional

## 📋 Documentation
- Comprehensive README with quick start guide
- Performance testing best practices
- Troubleshooting documentation
- CI/CD integration examples
- Advanced usage scenarios

## 🎯 Value Proposition
This enhancement demonstrates enterprise-level performance engineering practices and provides a solid foundation for continuous performance validation in the TypeScript API training material. It showcases:

- **Real-world Performance Testing**: Artillery.js integration with realistic scenarios
- **Progressive CI/CD**: Performance gates aligned with branch strictness
- **Monitoring & Observability**: Interactive dashboards and automated alerting
- **Regression Prevention**: Automated baseline comparison and threshold validation
- **Documentation Excellence**: Comprehensive guides for all skill levels

The implementation follows enterprise standards while maintaining educational value for training purposes.

🤖 Generated with [Claude Code](https://claude.ai/code)